Add local views
Unregistered users may create and edit "local" views now without
creating an account.  These views are stored in the DB as usual
presentations, and are accessed via a key stored in the cookies.

Local presentations may be recycled based on the last access time.  The
access time is set up in 1% of accesses only, to avoid excessive
payload.

This implements the first part of issue #45.