Add memcached for thread layouts
Threads' full HTML is cached.  It's on per-user, per-presentation basis,
so it uses much more memory than, perhaps, necessary.  I avoided making
it per-post as well (current posts are highlighted with a regexp and
CSS), but the rest remains.

This implements feature #18 for now.  After production testing we might
have to return to this feature.