Cache whole index pages for several seconds
This commit makes CachedThreads object provide an interface to query
individual threads from the cache for the whole index page.  It also
caches updates_at for the threads (which makes the actions that modify
the index respond slower than expected!).

The code does its best not to violate MVC, and passes a method to render
a thread from helper method only.